# specs/006-mettre-en-place/contracts/credits-api.openapi.yaml
openapi: 3.0.3
info:
  title: Fresquia Credit Management API
  description: API for managing credit packs, transactions, and user credit balances
  version: 1.0.0
  contact:
    name: Fresquia Development Team
servers:
  - url: https://api.fresquia.com
    description: Production server
  - url: https://staging-api.fresquia.com
    description: Staging server
paths:
...
  /api/payments/purchase-pack:
    post:
      summary: Purchase a credit pack
      description: Initiates the purchase of a credit pack through Stripe Checkout
      tags:
        - Payments
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchasePackRequest'
            examples:
              purchase_pro_pack:
                summary: Purchase Pro pack
                value:
                  pack_id: 2
                  success_url: "https://app.fresquia.com/billing/success"
                  cancel_url: "https://app.fresquia.com/billing/cancel"
      responses:
        '200':
          description: Purchase initiated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchasePackResponse'
              examples:
                checkout_created:
                  summary: Stripe Checkout session created
                  value:
                    checkout_url: "https://checkout.stripe.com/pay/cs_test_a1B2c3D4e5F6"
                    payment_intent_id: "pi_1J2k3l4m5n6o7p8q"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Credit pack not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/payments/webhook:
    post:
      summary: Handle Stripe webhook events
      description: Processes webhook events from Stripe for payment completion and other payment-related events
      tags:
        - Payments
      security: [] # Webhooks are authenticated via signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Stripe webhook event payload
      responses:
        '200':
          description: Webhook processed successfully
        '400':
          description: Invalid webhook signature or malformed payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    PurchasePackRequest:
      type: object
      properties:
        pack_id:
          type: integer
          description: ID of the credit pack to purchase
          example: 2
        success_url:
          type: string
          format: uri
          description: URL to redirect to after successful payment
          example: "https://app.fresquia.com/billing/success"
        cancel_url:
          type: string
          format: uri
          description: URL to redirect to after cancelled payment
          example: "https://app.fresquia.com/billing/cancel"
      required:
        - pack_id
        - success_url
        - cancel_url
    PurchasePackResponse:
      type: object
      properties:
        checkout_url:
          type: string
          format: uri
          description: Stripe Checkout URL for payment completion
          example: "https://checkout.stripe.com/pay/cs_test_a1B2c3D4e5F6"
        payment_intent_id:
          type: string
          description: Stripe Payment Intent ID
          example: "pi_1J2k3l4m5n6o7p8q"
      required:
        - checkout_url
        - payment_intent_id
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Not enough credits"
        code:
          type: string
          description: Error code
          example: "INSUFFICIENT_CREDITS"
        message:
          type: string
          description: Detailed error description
          example: "You need 5 more credits to complete this operation"
        details:
          type: object
          description: Additional error details
          example:
            required: 5
            available: 2.4
            operation: "image_generation"
            suggestion: "Purchase more credits to continue"
      required:
        - error
        - code
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalid_request:
              summary: Invalid request parameters
              value:
                error: "Invalid request"
                code: "INVALID_REQUEST"
                message: "The request parameters are invalid"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            auth_required:
              summary: Authentication required
              value:
                error: "Authentication required"
                code: "UNAUTHORIZED"
                message: "Please provide valid authentication credentials"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token for authentication
...